requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
--find-links https://data.pyg.org/whl/torch-2.8.0+cu126.html

hydra-core==1.3.2
hydra-colorlog
hydra-optuna-sweeper
omegaconf==2.3.0

torch
lightning====2.5.4
torch_ema==0.3
torch_geometric
torch-cluster
e3nn==0.5.6
tensorboard

basis-set-exchange==0.9.1
ase==3.22.1
numpy==1.26.4
scipy==1.12.0
pandas==2.2.1
plotly==5.18.0
pymatgen==2023.11.12
python-dotenv==1.0.1
rdkit
pyscf
lmdb==1.4.1
lz4==4.3.2
zarr
loguru
p_tqdm==1.4.0
tqdm==4.66.1
rootutils
mp-pyrho==0.3.0
ipdb==0.13.13
rich
pytest
pre-commit
ruff
setuptools<80.9